requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
pymysql==1.0.2
pytest==5.4.2
pandas==1.1.2
numpy==1.19.2
botocore== 1.15.32
boto3==1.12.32
s3fs==0.5.1
fsspec==0.8.0
scikit_learn==0.21.2
seaborn==0.10.1
matplotlib==3.2.2
requests==2.25.1
sportsipy==0.6.0
